parser remedy
Spent hours on Friday the 10th trying to understand why it was over
reading the first line of every section for instance the Mondays at 8:00
or 5:00. So I decided to go an alternate path and check for a Wednesday
class and verify it was a class not a lab by the length and duplicate
that class for a Monday.

AKA Steven, and Harrison check it. It works for my roomate and one of my
friends. Aurash feel free to check it anyway ;).